\name{about_bee}
\alias{about_bee}
\title{About the Berkeley Ecoinformatics Engine}
\usage{
  about_bee(as.df = FALSE)
}
\arguments{
  \item{as.df}{= FALSE Returns a list unless this set to
  \code{TRUE}}
}
\value{
  \code{list}
}
\description{
  About the Berkeley Ecoinformatics Engine
}
\examples{
\dontrun{
about_bee()
about_bee(as.df = TRUE)
}
}

