\docType{data}
\name{california_counties}
\alias{california_counties}
\title{california_counties}
\format{A data table with 58 rows and 3 variables}
\description{
  A data.frame containing list of California counties.
  Useful for most functions since the data is primarily
  state based.
}
\keyword{datasets}

