\docType{package}
\name{ecoengine}
\alias{ecoengine}
\alias{ecoengine-package}
\title{ecoengine}
\description{
ecoengine
}
\details{
The Berkeley Ecoengine \url{http://ecoengine.berkeley.edu}
provides an open API to a wealth of museum data contained
in the Berkeley natural history museums
url{https://bnhm.berkeley.edu/}. This R package provides a
programmatic interface to this rich repository of data
allowing for the data to be easily analyzed and visualized
or brought to bear in other contexts. This vignette
provides a brief overview of the package's capabilities.
}

