\name{ee_map}
\alias{ee_map}
\title{LeafletJS Map}
\usage{
ee_map(ee_obj, dest = tempdir(), title = "Ecoengine species map")
}
\arguments{
  \item{ee_obj}{A ecoengine object of type
  \code{observations} as generated by
  \code{\link{ee_observations}}}

  \item{dest}{Location where the html file and geojson file
  should be stored. Default is the temp directory}

  \item{title}{Title of the map.}
}
\description{
Builds an interactive map of locations for any species
}
\examples{
\dontrun{
lynx_data <- ee_observations(genus = "Lynx", georeferenced = TRUE, page = "all", quiet = TRUE)
ee_map(lynx_data)
}
}
\keyword{map}

