% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_map2.R
\name{ee_map2}
\alias{ee_map2}
\title{ee_map}
\usage{
ee_map2(ee_obj)
}
\arguments{
\item{ee_obj}{An input of class ecoengine}
}
\description{
Draws a leaflet map
}
\examples{
\dontrun{
chordata <- ee_observations(phylum = "chordata")
ee_map2(chordata)
}
}

