% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_utils.R
\name{ee_pages}
\alias{ee_pages}
\title{ee_pages - Returns total number of pages for any ecoengine request}
\usage{
ee_pages(ee, page_size = 1000)
}
\arguments{
\item{ee}{Object of class \code{ecoengine}}

\item{page_size}{Default page size. Currently set to \code{1000} package wide.}
}
\value{
integer
}
\description{
ee_pages - Returns total number of pages for any ecoengine request
}

