\name{ee_sensors}
\alias{ee_sensors}
\title{ee sensors}
\usage{
ee_sensors(...)
}
\arguments{
  \item{...}{same arguments as \code{\link{ee_sensors}}}
}
\value{
data.frame
}
\description{
Returns a full list of sensors with available data
}
\examples{
\dontrun{
full_sensor_list <- ee_sensors()
}
}

