\name{ee_sources}
\alias{ee_sources}
\title{Ecoengine data sources}
\usage{
ee_sources(foptions = list())
}
\arguments{
  \item{foptions}{A list of additional arguments. Currently
  this function takes none.}
}
\value{
\code{data.frame}
}
\description{
Returns a full list of data source supported by the
ecoengine
}
\examples{
\dontrun{
ee_sources()
}
}

