% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{full_sensor_list}
\alias{full_sensor_list}
\title{full_sensor_list}
\format{A data table with 40 rows and 13 variables. Contains units, variable, data_url, source, record, site_code,   method_name, begin_date, end_date, station_name, geojson.type, geojson.coordinates1, geojson.coordinates2}
\usage{
data(full_sensor_list)
}
\description{
A data.frame containing all the sensors currently available through the engine. This list can be updated anytime by calling the ee_sensors() function.
}
\keyword{datasets}

