\name{holos_checklists}
\alias{holos_checklists}
\title{<brief desc>}
\usage{
holos_checklists(subject = NULL, foptions = list())
}
\arguments{
  \item{subject}{Enter one of the following subjects:
  Mammals, Mosses, Beetles, Spiders, Amphibians, Ants,
  Fungi, Lichen, Plants.}

  \item{foptions}{= list() <what param does>}
}
\value{
data.frame
}
\description{
<full description>
}
\examples{
\dontrun{
mammals_list  <- holos_checklists(subject = "Mammals")
}
}

