\name{holos_sources}
\alias{holos_sources}
\title{Holos data sources}
\usage{
holos_sources(foptions = list())
}
\arguments{
  \item{foptions}{A list of additional arguments. Currently
  this function takes none.}
}
\value{
\code{data.frame}
}
\description{
Returns a full list of data source supported by the
ecoengine
}
\examples{
\dontrun{
holos_sources()
}
}

