\name{observations}
\alias{observations}
\title{Observations List}
\usage{
  observations(type = NULL)
}
\arguments{
  \item{type}{= NULL filter by type}
}
\value{
  \code{data.frame}
}
\description{
  API endpoint that represents a list of observations.
  Prameter options have not been fleshed out yet
}
\examples{
\dontrun{
observation()
}
}

