\docType{data}
\name{california_counties}
\alias{california_counties}
\title{california_counties}
\format{A data table with 58 rows and 3 variables. Contains county name, FIPS state code (06 in this case) and FIPS county code. See \href{http://en.wikipedia.org/wiki/FIPS_county_code}{Wikipedia} for more information.}
\usage{
data(california_counties)
}
\description{
A data.frame containing list of California counties. Useful
for most functions since the data is primarily state based.
}
\keyword{datasets}

