% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ee_checklists.R
\name{checklist_details}
\alias{checklist_details}
\title{Checklist details}
\usage{
checklist_details(list_name, ...)
}
\arguments{
\item{list_name}{URL of a checklist}

\item{...}{Additional arguments (currently not implemented)}
}
\value{
\code{data.frame}
}
\description{
Will return details on any checklist
}
\examples{
\dontrun{
spiders  <- ee_checklists(subject = "Spiders")
# Now retrieve all the details for each species on both lists
library(plyr)
spider_details <- ldply(spiders$url, checklist_details)
}
}
\seealso{
\code{\link{ee_checklists}}
}

