% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoengine_package.R
\docType{package}
\name{ecoengine}
\alias{ecoengine}
\alias{ecoengine-package}
\title{ecoengine}
\description{
\href{http://ecoengine.berkeley.edu}{The Berkeley Ecoengine} provides an open API to a wealth of museum data contained in the  \href{http://bnhm.berkeley.edu/}{Berkeley natural history museums}. This R package provides a programmatic interface to this rich repository of data allowing for the data to be easily analyzed and visualized or brought to bear in other contexts. This vignette provides a brief overview of the package's capabilities. The current version of the package works with API version 0.0.3
}
\details{
\href{http://ecoengine.berkeley.edu}{The Berkeley Ecoengine} provides an open API to a wealth of museum data contained in the  \href{http://bnhm.berkeley.edu/}{Berkeley natural history museums}. This R package provides a programmatic interface to this rich repository of data allowing for the data to be easily analyzed and visualized or brought to bear in other contexts. This vignette provides a brief overview of the package's capabilities. The current version of the package works with API version 0.0.3
}

