% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ee_checklists.R
\name{ee_checklists}
\alias{ee_checklists}
\title{Ecoengine checklists}
\usage{
ee_checklists(subject = NULL, foptions = list())
}
\arguments{
\item{subject}{Enter one of the following subjects: Mammals, Mosses, Beetles, Spiders, Amphibians, Ants, Fungi, Lichen, Plants.}

\item{foptions}{A list of additional arguments for \code{httr}. There is no reason to use this argument except for debugging purposes.}
}
\value{
data.frame
}
\description{
Retrieves exisitng checklists from the ecoengine database
}
\examples{
# all_lists  <- ee_checklists()
# mammals_list  <- ee_checklists(subject = "Mammals")
# spiders  <- ee_checklists(subject = "Spiders")
}

