% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ee_footprints}
\alias{ee_footprints}
\title{ee_footprints}
\usage{
ee_footprints(foptions = list())
}
\arguments{
\item{foptions}{A list of additional arguments for \code{httr}. There is no reason to use this argument except for debugging purposes.}
}
\value{
data.frame
}
\description{
List of ecoengine footprints.
}
\examples{
footprints <- ee_footprints()
}

