% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ee_sensors.R
\name{ee_sensors}
\alias{ee_sensors}
\title{ee_sensors}
\usage{
ee_sensors(page = NULL, page_size = 1000, remote_id = NULL,
  collection_code = NULL, source = NULL, min_date = NULL,
  max_date = NULL, foptions = list())
}
\arguments{
\item{page}{Page number}

\item{page_size}{The number of observations per page returned by the query. The package default is \code{25} observations. The API default, however, is \code{10}. Setting this number higher will require fewer calls (i.e. pages) but a large size may also fail or timeout.}

\item{remote_id}{Need to describe these parameters}

\item{collection_code}{Type of collection. Can be \code{CalAcademy}, \code{Private}, \code{VTM}, \code{CDFA}. \code{CalFlora} Others TBA}

\item{source}{Need to describe these parameters}

\item{min_date}{Lower date bound. Most be in format Y-m-D h:m:s. This package uses lubridate's \code{ymd_hms} function to format dates.}

\item{max_date}{upper date bound. Most be in format Y-m-D h:m:s. This package uses lubridate's \code{ymd_hms} function to format dates.}

\item{foptions}{A list of additional arguments for \code{httr}. There is no reason to use this argument except for debugging purposes.}
}
\description{
Returns UC reserve system sensor data
}
\examples{
# Currently there are only 40 sensors, so request only needs to be pages 1 and 2.
# ee_sensors()
# all_sensors <- ee_sensors()
}

