% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ee_sources}
\alias{ee_sources}
\title{Ecoengine data sources}
\usage{
ee_sources(foptions = list())
}
\arguments{
\item{foptions}{A list of additional arguments for \code{httr}. There is no reason to use this argument except for debugging purposes.}
}
\value{
\code{data.frame}
}
\description{
Returns a full list of data sources supported by the ecoengine
}
\examples{
source_list <- ee_sources()
}

