\name{holos_sensors}
\alias{holos_sensors}
\title{holos sensors}
\usage{
  holos_sensors(...)
}
\arguments{
  \item{...}{same arguments as \code{\link{holos_sensors}}}
}
\value{
  data.frame
}
\description{
  Returns a full list of sensors with available data
}
\examples{
\dontrun{
full_sensor_list <- holos_sensors()
}
}

